requirements.txt
certifi==2018.10.15
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0
django-ckeditor==5.6.1
django-js-asset==1.1.0
gunicorn==19.9.0
idna==2.7
Pillow==5.3.0
psycopg2==2.7.6.1
python-decouple==3.1
pytz==2018.7
requests==2.20.1
static3==0.7.0
urllib3==1.24.1
whitenoise==4.1.1
